\name{internals}
\alias{localnoise}
\alias{getpeakheights}
\alias{calculatebasis.gaussian}
\alias{calculatebasis.emg}
\alias{nnlslogbarrier}
\alias{linesearch}
\alias{postprocess.gaussian}
\alias{intermediate.gaussian}
\alias{gaussfun}
\alias{postprocess.emg}
\alias{erfc}
\alias{P1}
\alias{P2}
\alias{P3}
\alias{P3prime}
\alias{EMG}
\alias{determinemode}
\alias{dalpha}
\alias{dsigma}
\alias{dmu}
\alias{objective}
\alias{grad}
\alias{gridsearch}
\alias{fit.EMG}
\alias{objective.combine}
\alias{grad.combine}
\alias{fit.combine}
\alias{intermediateemg}
\alias{peakdetect}
\alias{fit.gauss}
\alias{formulacoef2function}
\alias{simplepeakdetect}
\alias{tableaveragine}
\alias{examples_boxes}
\alias{nnladlogbarrier}
\alias{linesearchlad}
\alias{calculatedenoisingbasis.gaussian}
\alias{calculatedenoisingbasis.emg}
\title{Internal functions}
\description{Functions not intended to be called directly be the user.}
