\name{myo500}
\alias{myo500}
\docType{data}
\title{Myoglobine dataset}

\usage{data(myo500)}

\description{A protein mass spectrum (MALDI) of Myoglobine at resolution
500fmol.}
\source{The data have kindly be provided by 
B. Gregorius and A. Tholey, Department of Experimental Medicine, Working
Group for Systematic Proteomics, Christian-Albrechts-Universitaet zu
Kiel. The dataset may be used in publications as long as this source
is quoted. 
}
%\references{
%}
\examples{
data(myo500)
plot(500)
}
\keyword{models}
