\name{read.mzXML}
\alias{read.mzXML}
\alias{new.mzXML}
\title{Import of data in mzXML format}
\description{The function has been extracted from the
  package \code{caMassClass}. It allows the import of
  data in mzXML format typically used for LC-MS data.}

\usage{
read.mzXML(filename) 
}
\arguments{
  \item{filename}{Path to the file that is to be imported.}  
}


\value{An object of class \code{mzXML}.}

\seealso{\link{analyzeLCMS}}
\keyword{models}
\examples{}
